refactor(contracts): combine staking.rs with node_registry.rs
